requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
markdown
pytz
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt
feedparser